Adding tests making the database stuff work

+++ Catamaran_Models/Interfaces/ITransactionModel.cs
@@ -7,7 +7,7 @@ namespace Catamaran_Models.Interfaces
     {
         public Guid TransactionId { get; set; }
         public DateTime TransactionDate { get; set; }
-        public long TransactionAmount { get; set; }
+        public decimal TransactionAmount { get; set; }
         public string Vendor { get; set; }
         public string Product { get; set; }
         public PaymentModes PaymentMethod { get; set; }
