+++ Catamaran_Models/Models/TransactionModel.cs
@@ -8,7 +8,7 @@ namespace Catamaran_Models.Models
     {
         public Guid TransactionId { get; set; }
         public DateTime TransactionDate { get; set; }
-        public long TransactionAmount { get; set; }
+        public decimal TransactionAmount { get; set; }
         public string Vendor { get; set; }
         public string Product { get; set; }
         public PaymentModes PaymentMethod { get; set; }
